kitchenai-community/src/kitchenai_community/llama_index_chat_engine_best_mode/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kitchenai_llama_index_chat_engine_best_mode"
version = "0.0.1"
authors = [{ name = "epuerta", email = "esteban_puerta@rhinosearch.io" }]
description = ''
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = []
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "django-ninja",
  "llama-index", 
  "chromadb", 
  "llama-index-vector-stores-chroma", 
  "boto3",
  "llama-index-llms-openai"
]

[project.urls]
Documentation = "https://github.com/epuerta/kitchenai_llama_index_chat_engine_best_mode#readme"
Issues = "https://github.com/epuerta/kitchenai_llama_index_chat_engine_best_mode/issues"
Source = "https://github.com/epuerta/kitchenai_llama_index_chat_engine_best_mode"

[tool.hatch.version]
path = "kitchenai_llama_index_chat_engine_best_mode/__about__.py"

[tool.falco.work]
server = "python manage.py migrate && python manage.py runserver {address}"
worker = "python manage.py qcluster"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/kitchenai_llama_index_chat_engine_best_mode tests}"

[tool.coverage.run]
source_pkgs = ["kitchenai_llama_index_chat_engine_best_mode", "tests"]
branch = true
parallel = true
omit = [
  "kitchenai_llama_index_chat_engine_best_mode/__about__.py",
]

[tool.coverage.paths]
kitchenai_llama_index_chat_engine_best_mode = ["kitchenai_llama_index_chat_engine_best_mode", "*/kitchenai_llama_index_chat_engine_best_mode/kitchenai_llama_index_chat_engine_best_mode"]
tests = ["tests", "*/kitchenai_llama_index_chat_engine_best_mode/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.envs.default.scripts]
docker-build = "docker build -t kitchenai_llama_index_chat_engine_best_mode ."


[tool.hatch.envs.dev]
dependencies = [
    "django-browser-reload",
    "django-debug-toolbar",
    "dj-notebook>=0.6.1",
    "Werkzeug[watchdog]>=3.0.1",
    "django-stubs[compatible-mypy]",
    "django-fastdev",
    "coverage[toml]>=6.5",
    "pre-commit",
    "pytest",
    "pytest-django",
    "pytest-sugar",
    "pytest-xdist",
    "falco-cli",
    "django-watchfiles",
    "git-cliff",
    "bump-my-version",
    "hatch-pip-compile",
]
lock-filename = "requirements-dev.txt"

[tool.hatch.envs.dev.scripts]
compile = "hatch-pip-compile --output-file requirements-dev.txt"